[GB.DB] * BUG: Correctly handle null dates.
[GB.DB.FORM]
* NEW: DataControl now uses a DateBox to edit date values.

[GB.FORM]
* NEW: DateBox handles null dates now.
* NEW: MaskBox hides its contents now when it is the default one and it 
  does not have the focus.
* NEW: DateChooser displays a little cursor following the mouse inside the
  calendar and the time panel.

[GB.GTK]
* BUG: Fix popup management.
* BUG: Editable ComboBox  controls now have the right background.
* NEW: Automatically temporarily release the keyboard and the mouse when 
  the debugger stops the program.


git-svn-id: svn://localhost/gambas/trunk@5066 867c0c6c-44f3-4631-809d-bfa615b0a4ec

+++ gb.qt4/src/main.cpp
@@ -1190,7 +1190,7 @@ int EXPORT GB_INFO(const char *key, void **value)
 extern Time	qt_x_time;
 #endif*/
 
-static void activate_main_window(int value)
+static void activate_main_window(intptr_t value)
 {
 	CWINDOW *active;
